Fix first block of CBC decryption; misc tidy up
fix a whoops in the CBC Oracle implementation by extending the struct with an
Option<AesKey> to store the IV used. This is crucial for CBC mode decryption
where the IV takes the place of the N-1 block ciphertext in the chained
decryption. Add a toy 2-16 round-trip implementation to demonstrate that it
works as expected.